+++ PlannR.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommandHandler.cs
@@ -31,6 +31,8 @@ namespace PlannR.Application.Features.Events.Commands.DeleteEvent
                 throw new NotFoundException(nameof(Event), request.EventId);
             }
 
+            if (!_authorisationService.CanAccessEntity(result)) throw new NotAuthorisedException();
+
             await _eventRepository.DeleteAsync(result);
 
             return Unit.Value;
