+++ PlannR.Application/Features/Routes/Commands/DeleteRoute/DeleteRouteCommandHandler.cs
@@ -31,6 +31,8 @@ namespace PlannR.Application.Features.Routes.Commands.DeleteRoute
                 throw new NotFoundException(nameof(Route), request.RouteId);
             }
 
+            if (!_authorisationService.CanAccessEntity(result)) throw new NotAuthorisedException();
+
             await _routeRepository.DeleteAsync(result);
 
             return Unit.Value;
