+++ PlannR.Application/Features/Transports/Commands/DeleteTransport/DeleteTransportCommandHandler.cs
@@ -31,6 +31,8 @@ namespace PlannR.Application.Features.Transports.Commands.DeleteTransport
                 throw new NotFoundException(nameof(Transport), request.TransportId);
             }
 
+            if (!_authorisationService.CanAccessEntity(result)) throw new NotAuthorisedException();
+
             await _transportRepository.DeleteAsync(result);
 
             return Unit.Value;
